Some cleanup

+++ src/lib/ipconv.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2009-2020, Peter Haag
+ *  Copyright (c) 2009-2022, Peter Haag
  *  Copyright (c) 2008-2008, SWITCH - Teleinformatikdienste fuer Lehre und Forschung
  *  All rights reserved.
  *
@@ -126,7 +126,7 @@ static int parse_ipv4(const char *src, uint32_t *dst, int *bytes) {
         const char *pch;
 
         if ((pch = strchr(digits, ch)) != NULL) {
-            unsigned int new = *tp * 10 + (pch - digits);
+            long new = *tp * 10 + (pch - digits);
 
             if (new > 255) return (0);
             if (!saw_digit) {
@@ -207,7 +207,7 @@ static int parse_ipv6(const char *src, uint64_t *dst, int *bytes) {
          * Since some memmove()'s erroneously fail to handle
          * overlapping regions, we'll do the shift by hand.
          */
-        const int n = tp - colonp;
+        const long n = tp - colonp;
         int i;
 
         for (i = 1; i <= n; i++) {
