+++ src/metric.c
@@ -298,7 +298,7 @@ __attribute__((noreturn)) void *MetricThread(void *arg) {
                 metric_chain = metric_chain->next;
                 LogInfo("Message sent for '%s', exporter: %d\n", metric_record->ident, exporterID);
             }
-            int ret = write(fd, message, offset);
+            ssize_t ret = write(fd, message, offset);
             if (ret < 0) {
                 LogError("write() error in %s line %d: %s", __FILE__, __LINE__, strerror(errno));
             }
